preprocessing/requirements.txt
numpy==1.26.4
Pillow==10.4.0
tqdm==4.66.4
torch==2.1.2+cpu
torchvision==0.16.2+cpu
torchaudio==2.1.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
faiss-cpu==1.8.0
git+https://github.com/openai/CLIP.git
